refactor(application): correctly setup application details page routing
